Moved EmployeeService to a specific directory. Changed the relationship between Employee and status in the Employee Entity from OneToOne to ManyToOne as this allows multiple employees to have the same status
