+++ employee-service/src/main/java/com/swiftedge/employeeservice/service/employee/EmployeeService.java
@@ -1,4 +1,4 @@
-package com.swiftedge.employeeservice.service;
+package com.swiftedge.employeeservice.service.employee;
 
 import com.swiftedge.employeeservice.dto.address.AddressRequestDTO;
 import com.swiftedge.employeeservice.dto.employee.EmployeeRequestDTO;
@@ -19,7 +19,6 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.reactive.function.client.WebClient;
 
 import java.util.List;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Collectors;
